requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
google-genai
edge-tts